refactor: Event와 EventPool 분리
