+++ src/event/EventPool.cpp
@@ -1,30 +1,10 @@
+#include "EventPool.hpp"
+
 #include <unistd.h>
 
-#include <cstring>
-#include <iostream>
-#include <new>
-#include <stdexcept>
-
-#include "event/event.hpp"
-#include "event/kqueue/kqueue.hpp"
+#include "kqueue/kqueue.hpp"
 #include "util/general/logging.hpp"
 #include "util/general/time.hpp"
-
-Event::Event(EventPool &pool_) : pool(pool_), flags() {}
-
-void Event::setReadEvent(const struct kevent &kev) {
-  this->kind = EventKind::kRead;
-  this->data = kev.data;
-  if (kev.flags & EV_EOF)
-    this->flags.set(EventFlag::kEOF);
-}
-
-void Event::setWriteEvent(const struct kevent &kev) {
-  this->kind = EventKind::kWrite;
-  this->data = kev.data;
-  if (kev.flags & EV_EOF)
-    this->flags.set(EventFlag::kEOF);
-}
 
 result_t::e EventPool::initKqueue() {
   fd_ = util::kqueue();
@@ -75,6 +55,7 @@ int16_t convert_to_evfilt(EventKind::e kind) {
       return EVFILT_WRITE;
 
     default:
+      util::debug_info("aborting due to unknown EventKind:", kind, false);
       abort();
   }
 }
