+++ src/server/Server.hpp
@@ -8,7 +8,8 @@
 #include <utility>
 
 #include "client/Client.hpp"
-#include "event/event.hpp"
+#include "event/Event.hpp"
+#include "event/EventPool.hpp"
 #include "util/FixedBuffer/FixedBuffer.hpp"
 #include "util/config/config.hpp"
 #include "util/general/result.hpp"
